+++ packages/discovery-components-react/.storybook/config.ts
@@ -1,5 +1,6 @@
 import { configure, addDecorator } from '@storybook/react';
 import { withInfo } from '@storybook/addon-info';
+import '@disco-widgets/styles';
 
 const req = require.context('../src', true, /\.stories.tsx$/);