+++ examples/discovery-components-react-example/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.11.0](https://github.ibm.com/Watson-Discovery/disco-widgets/compare/discovery-components-react-example@1.10.0...discovery-components-react-example@1.11.0) (2019-09-19)
+
+### Features
+
+- Refinements component basic ([#106](https://github.ibm.com/Watson-Discovery/disco-widgets/issues/106)) ([2416fc7](https://github.ibm.com/Watson-Discovery/disco-widgets/commit/2416fc7))
+
+# [1.10.0](https://github.ibm.com/Watson-Discovery/disco-widgets/compare/discovery-components-react-example@1.9.2...discovery-components-react-example@1.10.0) (2019-09-19)
+
+### Features
+
+- add discovery-styles package ([#109](https://github.ibm.com/Watson-Discovery/disco-widgets/issues/109)) ([7b819ac](https://github.ibm.com/Watson-Discovery/disco-widgets/commit/7b819ac))
+
 ## [1.9.2](https://github.ibm.com/Watson-Discovery/disco-widgets/compare/discovery-components-react-example@1.9.1...discovery-components-react-example@1.9.2) (2019-09-18)
 
 **Note:** Version bump only for package discovery-components-react-example