+++ packages/discovery-components-react/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.15.0](http:///discovery-components-react/compare/@disco-widgets/react-components@1.14.0...@disco-widgets/react-components@1.15.0) (2019-09-19)
+
+### Features
+
+- Refinements component basic ([#106](http:///discovery-components-react/issues/106)) ([2416fc7](http:///discovery-components-react/commits/2416fc7))
+
+# [1.14.0](http:///discovery-components-react/compare/@disco-widgets/react-components@1.13.8...@disco-widgets/react-components@1.14.0) (2019-09-19)
+
+### Features
+
+- add discovery-styles package ([#109](http:///discovery-components-react/issues/109)) ([7b819ac](http:///discovery-components-react/commits/7b819ac))
+
 ## [1.13.8](http:///discovery-components-react/compare/@disco-widgets/react-components@1.13.7...@disco-widgets/react-components@1.13.8) (2019-09-18)
 
 **Note:** Version bump only for package @disco-widgets/react-components